~move tests to the proper namespace
